Major code refactoring
