+++ src/_templates/launch_template.json
