+++ src/_templates/properties_template.json
